ENH: CIKLookup Improvements (#171)
* CLN: Start consolidating cik_lookup

* CLN: Continue moving

* CLN: flake8

* STRUCT: Move around files to flatten structure

* STRUCT: Continue moving cik_lookup

* CLN: Start cleaning up get_cik_map use

* FIX: Fix case sensitivity

* TEST: Add back test coverage for get_cik_map

* CLN: flake8

* DOC: Start updating docs

* CLN: Get rid of returns from validation helpers

* MAINT: Store normalized lookup in variable

* CLN: Get rid of unnecessary call to `get_ciks`

* MAINT: Rename _get_cik to _get_cik_from_html for clarity
