+++ docs/source/whatsnew.rst
@@ -6,6 +6,8 @@ What's New
 
 New features, bug fixes, and improvements for each release.
 
+.. include:: whatsnew/v0.3.2.rst
+
 .. include:: whatsnew/v0.3.1.rst
 
 .. include:: whatsnew/v0.3.0.rst
