+++ secedgar/exceptions.py
